+++ modele/classes/TypeRestaurant.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace bd\php\classes;
+namespace modele\classes\classes;
 class TypeRestaurant{
     private String $type;
